+++ lib/google_maps_url_extractor.dart
@@ -1,3 +1,5 @@
 library google_maps_url_extractor;
 
 export 'url_extractor.dart';
+export 'cache_manager.dart';
+export 'exceptions.dart';